Added Fake.Addresses.postalCode and a way to easily replace substring using multiple generators
