+++ Sources/Genything/gen/mutate/Gen+replacingOccurences.swift
@@ -25,4 +25,25 @@ public extension Gen {
             return result
         }
     }
+
+    /// Returns: A generator of strings which modify the `source` by replacing all occurence of `replace` strings their associated generator
+    ///
+    /// - Parameters:
+    ///   - source: The source or template `String` that will be modified
+    ///   - generators: Tuple of substring to replace with associated generator
+    ///
+    /// - Returns: The `String` generator
+    static func replacing(_ source: String, with generators: [(replace: String, by: Gen<String>)]) -> Gen<String> {
+        Gen<String> { ctx in
+            var result = source
+
+            generators.forEach { (substring, gen) in
+                while let rangeToReplace = result.range(of: substring) {
+                    result = result.replacingCharacters(in: rangeToReplace, with: gen.generate(context: ctx))
+                }
+            }
+
+            return result
+        }
+    }
 }
