+++ Tests/TrickeryTests/FakeAddressesTests.swift
@@ -1,0 +1,22 @@
+import XCTest
+@testable import Trickery
+
+class FakeAddressesTests: XCTestCase {
+    func test_postalCode() {
+        Fake.Addresses.postalCode.forEach { postalCode in
+            let subject = Array(postalCode)
+
+            XCTAssertEqual(7, subject.count)
+
+            XCTAssertTrue(subject[0].isLetter)
+            XCTAssertTrue(subject[1].isNumber)
+            XCTAssertTrue(subject[2].isLetter)
+
+            XCTAssertEqual(" ", subject[3])
+
+            XCTAssertTrue(subject[4].isNumber)
+            XCTAssertTrue(subject[5].isLetter)
+            XCTAssertTrue(subject[6].isNumber)
+        }
+    }
+}
